+++ CHANGELOG.md
@@ -46,6 +46,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Manuel Puyol*
 
+* Linters can autocorrect cases with custom classes.
+
+    *Manuel Puyol*
+
 ## 0.0.51
 
 ### Breaking changes